requirements.txt
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
black==24.10.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.6
cryptography==45.0.6
dill==0.4.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.2
idna==3.10
iniconfig==2.1.0
isort==5.13.2
mccabe==0.7.0
mpmath==1.3.0
mypy==1.13.0
mypy_extensions==1.1.0
numpy==2.2.0
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.4.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.2
pyspnego==0.11.2
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qiskit==2.1.2
qiskit-ibm-runtime==0.41.1
requests==2.32.5
requests_ntlm==1.3.0
rsa==4.9.1
rustworkx==0.17.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
stevedore==5.5.0
symengine==0.13.0
sympy==1.14.0
tenacity==9.1.2
tomlkit==0.13.3
types-Pillow==10.2.0.20240822
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
